# Improve performance of element handling