@baoilleach 2017-07-24
Updated timings, for the record, with 3 signif figures and 10x more iterations:

Timing comparison: get the symbol for elements 1 to 100 and convert back to atomic number (repeat x100000). 8.05s old vs 0.187s new. A more realistic conversion is to just do carbon: 1.34s old vs 0.078s new.