@ghutchis 2017-08-04
Yep, it looks good. Sorry for the delay