@n-yoshikawa 2019-06-09
Could you show me how to use `OBElements` in Python binding?
For example, how can I replace the following code?
```python
import openbabel

etab = openbabel.OBElementTable()
print(etab.GetSymbol(6))
```