@baoilleach 2019-06-09
https://open-babel.readthedocs.io/en/latest/UseTheLibrary/migration.html#handling-of-elements-and-related-information

On Sun, 9 Jun 2019, 09:40 Naruki Yoshikawa, <notifications@github.com>
wrote:

> Could you show me how to use OBElements in Python binding?
> For example, how can I replace the following code?
>
> import openbabel
>
> etab = openbabel.OBElementTable()print(etab.GetSymbol(6))
>
> —
> You are receiving this because you authored the thread.
> Reply to this email directly, view it on GitHub
> <https://github.com/openbabel/openbabel/pull/1601?email_source=notifications&email_token=AAALCGIO35247YSN4P4PXBTPZS6YVA5CNFSM4DS2IKEKYY3PNVWWK3TUL52HS4DFVREXG43VMVBW63LNMVXHJKTDN5WW2ZLOORPWSZGODXIGFSY#issuecomment-500196043>,
> or mute the thread
> <https://github.com/notifications/unsubscribe-auth/AAALCGOZ5ZS4BQMBAPYU5QDPZS6YVANCNFSM4DS2IKEA>
> .
>