@n-yoshikawa 2019-06-09
I tried the following `test.py` in the current master,
```
from openbabel import openbabel
  
print(openbabel.OBElements.GetSymbol(6))
```
but it yielded the followng error.
```
$ python3 test.py 
Traceback (most recent call last):
  File "test.py", line 3, in <module>
    print(openbabel.OBElements.GetSymbol(6))
AttributeError: module 'openbabel.openbabel' has no attribute 'OBElements'
```