@baoilleach 2019-06-09
Just GetSymbol will work without the OBElements.

On Sun, 9 Jun 2019, 10:30 Naruki Yoshikawa, <notifications@github.com>
wrote:

> I tried the following test.py in the current master,
>
> from openbabel import openbabel
>
> print(openbabel.OBElements.GetSymbol(6))
>
> but it yielded the followng error.
>
> $ python3 test.py
> Traceback (most recent call last):
>   File "test.py", line 3, in <module>
>     print(openbabel.OBElements.GetSymbol(6))
> AttributeError: module 'openbabel.openbabel' has no attribute 'OBElements'
>
> —
> You are receiving this because you authored the thread.
> Reply to this email directly, view it on GitHub
> <https://github.com/openbabel/openbabel/pull/1601?email_source=notifications&email_token=AAALCGIMJBQYZF7WH4HAFBLPZTETHA5CNFSM4DS2IKEKYY3PNVWWK3TUL52HS4DFVREXG43VMVBW63LNMVXHJKTDN5WW2ZLOORPWSZGODXIGZPI#issuecomment-500198589>,
> or mute the thread
> <https://github.com/notifications/unsubscribe-auth/AAALCGODS2U3OMT37K6KUVTPZTETHANCNFSM4DS2IKEA>
> .
>